+++ examples/01-message-codec/Main.kt
@@ -2,12 +2,11 @@ package examples.messagecodec
 
 import io.github.termtate.kotlinosc.codec.decodeFromByteArray
 import io.github.termtate.kotlinosc.codec.encodeToByteArray
-import io.github.termtate.kotlinosc.type.OscMessage
 import io.github.termtate.kotlinosc.type.OscPacket
-import io.github.termtate.kotlinosc.type.invoke
+import io.github.termtate.kotlinosc.type.oscMessageOf
 
 fun main() {
-    val message = OscMessage("/synth/freq", 440, 0.75f, "lead")
+    val message = oscMessageOf("/synth/freq", 440, 0.75f, "lead")
     val payload = message.encodeToByteArray()
     val decoded = OscPacket.decodeFromByteArray(payload)