+++ examples/03-udp-client-send/Main.kt
@@ -1,15 +1,14 @@
 package examples.udpclientsend
 
 import io.github.termtate.kotlinosc.transport.dsl.oscClient
-import io.github.termtate.kotlinosc.type.OscMessage
-import io.github.termtate.kotlinosc.type.invoke
+import io.github.termtate.kotlinosc.type.oscMessageOf
 import kotlinx.coroutines.runBlocking
 
 fun main(): Unit = runBlocking {
     val client = oscClient("127.0.0.1", 9000)
 
     try {
-        val message = OscMessage.invoke("/demo/ping", "hello from kotlinosc")
+        val message = oscMessageOf("/demo/ping", "hello from kotlinosc")
         client.send(message)
         println("Sent message to 127.0.0.1:9000 -> $message")
     } finally {